+++ polaris/evaluate/metrics/__init__.py
@@ -1,7 +1,7 @@
-from polaris.evaluate.metrics.generic_metrics import (
-    cohen_kappa_score,
-    absolute_average_fold_error,
-    spearman,
-    pearsonr,
-)
 from polaris.evaluate.metrics.docking_metrics import rmsd_coverage
+from polaris.evaluate.metrics.generic_metrics import (
+    absolute_average_fold_error,
+    cohen_kappa_score,
+    pearsonr,
+    spearman,
+)
